+++ src/OpenTelemetry.Exporter.Prometheus.AspNetCore/CHANGELOG.md
@@ -24,8 +24,8 @@ Notes](../../RELEASENOTES.md).
   correctly during Prometheus serialization.
   ([#7184](https://github.com/open-telemetry/opentelemetry-dotnet/issues/7184))
 
-* Use the canonical representation for "quantile" and "le" label values when
-  using OpenMetrics.
+* Use the canonical representation for histogram "le" label values when using
+  OpenMetrics.
   ([#7218](https://github.com/open-telemetry/opentelemetry-dotnet/issues/7218))
 
 ## 1.15.3-beta.1